requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
aiosqlite==0.20.0
greenlet==3.0.3
